Fix RemainingCapacity to use mAh instead of percentage
Critical bug fix: RemainingCapacity (HID usage 0x66) should contain
absolute mAh values, not percentage. macOS was calculating percentage
as RemainingCapacity/FullChargeCapacity = 100/7200 = 1.4%, which is
why it displayed "UPS 1%".

Changes:
- Changed RemainingCapacity from 8-bit (0-255) to 16-bit (0-65535) in HID descriptor
- Updated UPS_BatteryStateTypeDef.remaining_capacity from uint8_t to uint16_t
- Changed default value from 100 (percent) to 7200 (mAh)
- Button press now sets 360 mAh (5%) or 7200 mAh (100%)
- Updated BuildUPSReport to send 16-bit mAh value (14 bytes total, was 13)
- Fixed low battery threshold check to use mAh (10% of full_charge_capacity)
- All byte offsets in report shifted by 1 to accommodate 16-bit RemainingCapacity

This fix should make macOS properly calculate battery percentage:
360/7200 = 5% and 7200/7200 = 100%

+++ Projects/NUCLEO-C071RB/Applications/USBX/Ux_Device_HID_Standalone/USBX/App/ux_device_ups.h
@@ -44,7 +44,7 @@ typedef struct
   uint8_t below_capacity_limit;    /* Below capacity limit flag */
   uint8_t capacity_mode;           /* Capacity mode */
   uint8_t rechargeable;            /* Battery is rechargeable */
-  uint8_t remaining_capacity;      /* Remaining battery capacity (0-100%) */
+  uint16_t remaining_capacity;     /* Remaining battery capacity in mAh */
   uint16_t full_charge_capacity;   /* Full charge capacity in mAh */
   uint16_t design_capacity;        /* Design capacity in mAh */
   uint16_t voltage;                /* Present voltage in mV */
